I noticed that formating the file "pkg/front_end/lib/src/fasta/fasta_codes_cfe_generated.dart" from the dart sdk was quite slow, took a look into it and made some changes.
It's still quite slow, but at least it's better.

I've:

- Updated _compareScore to "cache" the costs and inlined the `compareTo` call (the sdk version takes a num and don't know it's already been compared to 0).
- Replaced the usage of Set in a few places with markings on the actual classes.

My measured runtimes:

```
AS-IS
Formatted 1 file (0 changed) in 3.53 seconds.
Formatted 1 file (0 changed) in 3.65 seconds.
Formatted 1 file (0 changed) in 3.64 seconds.
Formatted 1 file (0 changed) in 3.70 seconds.
Formatted 1 file (0 changed) in 3.85 seconds.
Formatted 1 file (0 changed) in 3.70 seconds.
Formatted 1 file (0 changed) in 3.76 seconds.
Formatted 1 file (0 changed) in 3.59 seconds.
Formatted 1 file (0 changed) in 3.77 seconds.
Formatted 1 file (0 changed) in 3.80 seconds.

_compareScore change
Formatted 1 file (0 changed) in 3.21 seconds.
Formatted 1 file (0 changed) in 3.35 seconds.
Formatted 1 file (0 changed) in 3.28 seconds.
Formatted 1 file (0 changed) in 3.10 seconds.
Formatted 1 file (0 changed) in 3.06 seconds.
Formatted 1 file (0 changed) in 3.35 seconds.
Formatted 1 file (0 changed) in 3.39 seconds.
Formatted 1 file (0 changed) in 3.08 seconds.
Formatted 1 file (0 changed) in 3.15 seconds.
Formatted 1 file (0 changed) in 3.31 seconds.

Difference at 95.0% confidence
        -0.471 +/- 0.105388
        -12.7332% +/- 2.84909%
        (Student's t, pooled s = 0.112163)

Mark Spans instead of using set in _calculateCost
Formatted 1 file (0 changed) in 2.96 seconds.
Formatted 1 file (0 changed) in 2.92 seconds.
Formatted 1 file (0 changed) in 2.98 seconds.
Formatted 1 file (0 changed) in 2.99 seconds.
Formatted 1 file (0 changed) in 3.10 seconds.
Formatted 1 file (0 changed) in 3.20 seconds.
Formatted 1 file (0 changed) in 2.93 seconds.
Formatted 1 file (0 changed) in 2.84 seconds.
Formatted 1 file (0 changed) in 3.09 seconds.
Formatted 1 file (0 changed) in 3.05 seconds.

Difference at 95.0% confidence
        -0.222 +/- 0.107951
        -6.87732% +/- 3.34422%
        (Student's t, pooled s = 0.114891)

Mark NestingLevels instead of using set in _calculateSplits
Formatted 1 file (0 changed) in 2.77 seconds.
Formatted 1 file (0 changed) in 2.97 seconds.
Formatted 1 file (0 changed) in 2.86 seconds.
Formatted 1 file (0 changed) in 2.90 seconds.
Formatted 1 file (0 changed) in 3.01 seconds.
Formatted 1 file (0 changed) in 2.83 seconds.
Formatted 1 file (0 changed) in 2.88 seconds.
Formatted 1 file (0 changed) in 2.88 seconds.
Formatted 1 file (0 changed) in 2.88 seconds.
Formatted 1 file (0 changed) in 2.90 seconds.

Difference at 95.0% confidence
        -0.118 +/- 0.0826868
        -3.92548% +/- 2.75073%
        (Student's t, pooled s = 0.0880025)
```

Which is a combined changed of
```
Difference at 95.0% confidence
        -0.811 +/- 0.079311
        -21.9248% +/- 2.14412%
        (Student's t, pooled s = 0.0844097)
```

i.e ~800 ms or almost 22% faster on this - apparently particularly bad - file.